requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
docker>=7.0.0
httpx>=0.26.0
PyJWT>=2.8.0
logtail-python>=0.3.0
